chore: move rT to solution
